+++ EasyAccess/EasyAccess.cs
@@ -34,7 +34,9 @@ public class EasyAccess : Mod
                 // ignored
             }
 
-            return this._config = config ?? new ModConfig();
+            this._config = config ?? new ModConfig();
+            Log.Trace(this._config.ToString());
+            return this._config;
         }
     }
 
@@ -43,6 +45,11 @@ public class EasyAccess : Mod
     {
         I18n.Init(helper.Translation);
         Log.Monitor = this.Monitor;
+
+        if (this.Helper.ModRegistry.IsLoaded("furyx639.FuryCore"))
+        {
+            Log.Alert("Remove FuryCore, it is no longer needed by this mod!");
+        }
 
         // Events
         this.Helper.Events.Content.AssetRequested += this.OnAssetRequested;