+++ TooManyAnimals/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Too Many Animals Change Log
 
-## 1.3.0 (July 1, 2022)
+## 2.0.1 (Unreleased)
+
+* Added alert if FuryCore is installed.
+* Log config options for debugging.
+
+## 2.0.0 (July 1, 2022)
 
 * Updated to SMAPI 3.15.0.
 * Removed dependency on FuryCore.