+++ ToolbarIcons/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Toolbar Icons Change Log
 
-## 1.0.0 (July 1, 2022)
+## 2.0.1 (Unreleased)
+
+* Added alert if FuryCore is installed.
+
+## 2.0.0 (July 2, 2022)
+
+* Bumped version number to avoid conflict on Toolbar Icons mod page.
+
+## 1.0.0 (Unreleased)
 
 * Initial Version